refactor(lotto): rename domain to model and some methods
리팩토링 : MVC 구조 및 이해의 편의를 위해 이름 변경
	- domain 패키지명을 model로 변경
	- validator의 메소드명을 isValidPrice에서 isThousandMultiples로 변경
