+++ src/main/java/lotto/model/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model;
 
 import java.util.ArrayList;
 import java.util.List;
